third_party/WebKit/LayoutTests/http/tests/media/media-source/mediasource-htmlmediaelement-lifetime-expected.txt

sourceOpened called.
Running the garbage collector.
Setting MediaSource duration.
durationChanged called.

PASS Tests that the MediaSource keeps the HTMLMediaElement alive. 
